+++ frontend/app/.server/domain/services/mock-data.ts
